validating user when performing crud on beer

+++ src/main/java/com/example/spring_demo/controllers/BeerMvcController.java
@@ -2,12 +2,14 @@ package com.example.spring_demo.controllers;
 
 import com.example.spring_demo.exceptions.DuplicateEntityException;
 import com.example.spring_demo.exceptions.EntityNotFoundException;
+import com.example.spring_demo.helpers.AuthenticationHelper;
 import com.example.spring_demo.helpers.BeerMapper;
 import com.example.spring_demo.models.*;
 import com.example.spring_demo.services.BeerService;
 import com.example.spring_demo.services.StyleService;
 import com.example.spring_demo.services.UserService;
 import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -25,12 +27,14 @@ public class BeerMvcController {
     private final UserService userService;
     private final StyleService styleService;
     private final BeerMapper beerMapper;
+    private final AuthenticationHelper authenticationHelper;
 
-    public BeerMvcController(BeerService beerService, UserService userService, StyleService styleService, BeerMapper beerMapper) {
+    public BeerMvcController(BeerService beerService, UserService userService, StyleService styleService, BeerMapper beerMapper, AuthenticationHelper authenticationHelper) {
         this.beerService = beerService;
         this.userService = userService;
         this.styleService = styleService;
         this.beerMapper = beerMapper;
+        this.authenticationHelper = authenticationHelper;
     }
 
     @GetMapping
@@ -44,7 +48,7 @@ public class BeerMvcController {
     }
 
     @GetMapping("/{id}")
-    public ModelAndView getBeerById(@PathVariable int id) {
+    public ModelAndView getBeerById(@PathVariable int id, HttpSession session) {
         ModelAndView modelAndView = new ModelAndView();
         try {
             Beer beer = beerService.getBeerById(id);
@@ -59,18 +63,23 @@ public class BeerMvcController {
     }
 
     @GetMapping("/new")
-    public String createFormBeer(Model model) {
+    public String createFormBeer(Model model, HttpSession session) {
+
+        if (getCurrentUser(session).getUsername().equals("Anonymous User")) {
+            return "redirect:/auth/login";
+        }
+
         model.addAttribute("beer", new BeerDto());
         return "CreateBeer";
     }
 
     @PostMapping("/new")
-    public String saveBeer(@Valid @ModelAttribute("beer") BeerDto beerDto, BindingResult errors) {
+    public String saveBeer(@Valid @ModelAttribute("beer") BeerDto beerDto, BindingResult errors, HttpSession session) {
         if (errors.hasErrors()) {
             return "CreateBeer";
         }
         try {
-            User user = userService.getById(3);
+            User user = authenticationHelper.tryGetUser(session);
             Beer beer = beerMapper.fromDto(beerDto);
 
             beerService.create(beer, user);
@@ -83,9 +92,17 @@ public class BeerMvcController {
 
 
     @GetMapping("/{id}/update")
-    public String getUpdateBeerPage(@PathVariable int id, Model model) {
+    public String getUpdateBeerPage(@PathVariable int id, Model model, HttpSession session) {
+        if (getCurrentUser(session).getUsername().equals("Anonymous User")) {
+            return "redirect:/auth/login";
+        }
         try {
+            User user = getCurrentUser(session);
             Beer beer = beerService.getBeerById(id);
+            if (!user.isAdmin() && !user.getUsername().equals(beer.getCreatedBy().getUsername())) {
+                return "Forbidden";
+            }
+
             BeerDto beerDto = beerMapper.toDto(beer);
             model.addAttribute("beer", beerDto);
             return "UpdateBeer";
@@ -94,6 +111,7 @@ public class BeerMvcController {
             return "Not-found";
         }
     }
+
 
     @PostMapping("/{id}/update")
     public String updateBeer(@Valid @ModelAttribute("beer") BeerDto beerDto, BindingResult errors, @PathVariable int id, Model model) {
@@ -113,9 +131,17 @@ public class BeerMvcController {
 
 
     @GetMapping("/{id}/delete")
-    public String deleteBeer(@PathVariable int id, Model model) {
+    public String deleteBeer(@PathVariable int id, Model model, HttpSession session) {
+        if (getCurrentUser(session).getUsername().equals("Anonymous User")) {
+            return "redirect:/auth/login";
+        }
         try {
-            User user = userService.getById(3);
+            User user = getCurrentUser(session);
+            Beer beer = beerService.getBeerById(id);
+            if (!user.isAdmin() && !user.getUsername().equals(beer.getCreatedBy().getUsername())) {
+                return "Forbidden";
+            }
+
             beerService.delete(id, user);
             return "redirect:/mvc/beers";
         } catch (EntityNotFoundException e) {
@@ -137,5 +163,18 @@ public class BeerMvcController {
     @ModelAttribute("requestURI")
     public String requestURI(final HttpServletRequest request) {
         return request.getRequestURI();
+    }
+
+    @ModelAttribute("user")
+    public User getCurrentUser(HttpSession session) {
+        User user;
+        try {
+            user = authenticationHelper.tryGetUser(session);
+        } catch (Exception e) {
+            user = new User();
+            user.setAdmin(false);
+            user.setUsername("Anonymous User");
+        }
+        return user;
     }
 }
